service/requirements.txt
# Web API
fastapi==0.133.0
uvicorn[standard]==0.30.6

# Settings / parsing
pydantic==2.10.6
pydantic-settings==2.7.1

requests==2.32.5
PyYAML==6.0.2
python-dateutil==2.9.0.post0

# LangChain ecosystem
langgraph==0.2.62
langchain-text-splitters==1.1.1

# DB drivers (your db_tool supports both)
psycopg[binary]==3.2.5
psycopg2-binary==2.9.11

# File parsing
PyMuPDF==1.24.14
pdfplumber==0.11.5
pypdf==6.7.3
openpyxl==3.1.5
Pillow==10.4.0

# Office docs
python-pptx==0.6.23
python-docx==1.1.2

# Google Drive
google-api-python-client==2.160.0
google-auth==2.38.0